+++ dkinst/__init__.py
@@ -1,4 +1,4 @@
 """Den K Simple Installer"""
 
 __author__ = "Den Kras"
-__version__ = '0.22.1'
+__version__ = '0.23.2'